+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 2.10.3 (2026-05-26)
+
+- Fix S3 uploads ([#428](https://github.com/datagouv/hydra/pull/428))
+- One bucket per env, prefixes made from file extentions ([#429](https://github.com/datagouv/hydra/pull/429))
+
+
+## 2.10.2 (2026-05-26)
+
+- docs: fix minor enqueue docstring error
+- feat(s3): S3 Object Storage via boto3 ([#415](https://github.com/datagouv/hydra/pull/415))
+
+
 ## 2.10.1 (2026-05-19)
 
 - feat: db only parquet export ([#424](https://github.com/datagouv/hydra/pull/424))